Move cuda_device_id to it's own header.
